Added muzzle flash effect Added death explosion effect Added effect when projectile hits character Added character knockback when hit Increased default size of projectiles 2x Fixed warnings in weapon and upgrade Minor refactoring work

+++ Project/src/game/objects/Upgrade.cpp
@@ -108,7 +108,7 @@ bool Upgrade::multiActive()
 	return m_extraProj.active;
 }
 
-float Upgrade::multiCount()
+unsigned int Upgrade::multiCount()
 {
 	return m_extraProj.nr;
 }
